만남 발자국 시스템 구현

+++ src/main/java/com/kslj/mannam/domain/user/dto/UserResponseDto.java
@@ -21,7 +21,7 @@ public class UserResponseDto {
     Long age;
     String personalities;
     String interests;
-    float footprint;
+    Double footprint;
     String question;
     String questionList;
     String imgUrl;
